+++ packages/immutable/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @solid-primitives/immutable
 
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [6680ab9]
+  - @solid-primitives/utils@6.4.0
+  - @solid-primitives/keyed@1.5.3
+
 ## 1.1.3
 
 ### Patch Changes